cube: rename cube-domE to cube-desktop
The cube-domE name is confusing, and we constantly have to explain what
the "E" stands for.

Also, as the design progressed, domE is not an essential part of the
system .. it is optional. As such, it shouldn't be designated as a
"domain".

By dropping the domE and making it cube-desktop, we resolve both the
confusion and the inconsistency.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>