overc-conftools: Add puppet infrastructure for assigning wildcard selected network devices
In order to generate images which will properly boot on hardware with
various network devices names it is desireable for some configurations
to own all the physical network devices specified by wild cards.

The additional specification in the lxc config file is:

overc.network.link = all
overc.network.linkdevs = eth* wl* en*

When the overc.network.link is set to all, the fixed or wild card
network devices can all be assigned to the container when the puppet
scripts initialize the first time.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>