lxc: Fix lxc-attach to output the pty stream after EPOLLHUP
The pty output may not get flushed on a short lived process which has
a very small output buffer.  Example:

   lxc-attach -n cube-desktop cat /etc/hostname

OR
   hst=`lxc-attach -n cube-desktop cat /etc/hostname`
   echo $hst

What will happen is that no output will be received to the controlling
pty.  When an EPOLLHUP is received the errors should not be printed
for the reads and writes, just process output and call it a day.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>