meta-overc: reorder the network device list
If more than 2 physical network interfaces were
assigned to a container, such as "eth0", "eth1"
and "eth2", and when this container stop, these
three network interfaces would return to essential
system with interfaces named changed to "eth1", "eth2"
and "dev5".

If the order of the interfaces were reverted to assinged
to container, when the container stop and reture those
interfaces to essential system the kernel would not
change those interfaces name.

Signed-off-by: fli <fupan.li@windriver.com>