cube-update: Fix the cubepkgcheck parse newer packages error
If the system network unreachable or smart channel could not work
as expected, it should display some WARNING information when users
login the system.

Create a smart_newer.sh script to deal with the "smart newer" pacakges
updating information. It could also be used by some applications. Such as
overc newer also need the similar function to find the newer packages in
the different containers or host.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>