cube-cmd-server: Allow cube-console to persist through a cube-cmd-server restart
In dom0 the cgexec command can be used to start the cube-console
outside the cube-cmd-server cgroup which will allow it to persist when
the cube-cmd-server is restarted.

This fixes the long standing work around where you have to run
"cube-console dom0" twice in order to perform an operation like
"cube-ctl add ...".

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>