overc-system-agent/run_server.py: Add basic auth and optional bind address
The overc agent rest api server needs to have basic authentication on
at all times.  The default user name and password are in the script
(adm/adm), but it is expected that an end use of the rest api server
will change the user name and password.

The addition of the basic auth method has pulled in two additional
depencies which are python-passlib and python-flask-httpauth.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>